Updated README.md (see commit description)
Reworded the installing section to only describe how to install as a jar file, thus removing the section about installing as software. This was done as I have disided to change the direction that I want to take the project in. 

Fixed the odd looking interatction between the jfr logo licence image and it's footnote marker. I should look less silly. 😛 

Fixed a few typos here and there.